+++ packages/contracts/foundry_test/CDPManager.redemptions.t.sol
@@ -8,8 +8,8 @@ import {eBTCBaseInvariants} from "./BaseInvariants.sol";
 contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
     // Storage array of cdpIDs when impossible to calculate array size
     bytes32[] cdpIds;
-    uint public mintAmount = 1e18;
-    uint private ICR_COMPARE_TOLERANCE = 1000000; //in the scale of 1e18
+    uint256 public mintAmount = 1e18;
+    uint256 private ICR_COMPARE_TOLERANCE = 1000000; //in the scale of 1e18
     address payable[] users;
 
     function setUp() public override {
@@ -20,8 +20,8 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
     }
 
     function testCDPManagerSetMinuteDecayFactorDecaysBaseRate() public {
-        uint newMinuteDecayFactor = (500 + 999037758833783000);
-        uint timePassed = 600; // seconds/60 => minute
+        uint256 newMinuteDecayFactor = (500 + 999037758833783000);
+        uint256 timePassed = 600; // seconds/60 => minute
 
         address user = _utils.getNextUserAddress();
 
@@ -31,7 +31,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
 
         collateral.approve(address(borrowerOperations), type(uint256).max);
 
-        uint debt = 2e17;
+        uint256 debt = 2e17;
 
         console.log("debt %s", debt);
 
@@ -48,8 +48,8 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         // Set the initial baseRate to a non-zero value via rdemption
         console.log("balance: %s", eBTCToken.balanceOf(user));
         eBTCToken.approve(address(cdpManager), type(uint256).max);
-        uint _redeemDebt = 1;
-        (bytes32 firstRedemptionHint, uint partialRedemptionHintNICR, , ) = hintHelpers
+        uint256 _redeemDebt = 1;
+        (bytes32 firstRedemptionHint, uint256 partialRedemptionHintNICR, , ) = hintHelpers
             .getRedemptionHints(_redeemDebt, (priceFeedMock.fetchPrice()), 0);
         cdpManager.redeemCollateral(
             _redeemDebt,
@@ -61,16 +61,16 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
             1e18
         );
 
-        uint initialRate = cdpManager.baseRate();
+        uint256 initialRate = cdpManager.baseRate();
 
         console.log("baseRate: %s", cdpManager.baseRate());
 
         // Calculate the expected decayed base rate
-        uint decayFactor = cdpManager.minuteDecayFactor();
+        uint256 decayFactor = cdpManager.minuteDecayFactor();
         console.log("decayFactor: %s", decayFactor);
-        uint _decayMultiplier = _decPow(decayFactor, (timePassed / 60));
+        uint256 _decayMultiplier = _decPow(decayFactor, (timePassed / 60));
         console.log("_decayMultiplier: %s", _decayMultiplier);
-        uint expectedDecayedBaseRate = (initialRate * _decayMultiplier) /
+        uint256 expectedDecayedBaseRate = (initialRate * _decayMultiplier) /
             cdpManager.DECIMAL_PRECISION();
 
         // Fast forward time by 1 minute
@@ -84,18 +84,18 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         vm.stopPrank();
     }
 
-    function testMultipleRedemption(uint _cdpNumber, uint _collAmt) public {
+    function testMultipleRedemption(uint256 _cdpNumber, uint256 _collAmt) public {
         vm.assume(_cdpNumber > 1);
         vm.assume(_cdpNumber <= 1000);
         vm.assume(_collAmt > 22e17);
         vm.assume(_collAmt <= 10000e18);
-        uint _price = priceFeedMock.getPrice();
+        uint256 _price = priceFeedMock.getPrice();
 
         // open random cdps with increasing ICR
         address payable[] memory _borrowers = _utils.createUsers(_cdpNumber + 1);
         bytes32[] memory _cdpIds = new bytes32[](_cdpNumber);
-        for (uint i = 1; i <= _cdpNumber; ++i) {
-            uint _debt = _utils.calculateBorrowAmount(
+        for (uint256 i = 1; i <= _cdpNumber; ++i) {
+            uint256 _debt = _utils.calculateBorrowAmount(
                 _collAmt,
                 _price,
                 (COLLATERAL_RATIO + (i * 5e15))
@@ -104,8 +104,8 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
             bytes32 _cdpId = _openTestCDP(_borrowers[i], _collAmt, _debt);
             _cdpIds[i - 1] = _cdpId;
             if (i > 1) {
-                uint _icr = cdpManager.getICR(_cdpId, _price);
-                uint _prevICR = cdpManager.getICR(_cdpIds[i - 2], _price);
+                uint256 _icr = cdpManager.getICR(_cdpId, _price);
+                uint256 _prevICR = cdpManager.getICR(_cdpIds[i - 2], _price);
                 require(_icr > _prevICR, "!icr");
                 require(_icr > CCR, "!icr>ccr");
             }
@@ -115,28 +115,28 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
 
         // prepare redemption by picking a random number of CDPs to redeem
         address _redeemer = _borrowers[0];
-        uint _debt = _utils.calculateBorrowAmount(_collAmt, _price, COLLATERAL_RATIO * 1000);
+        uint256 _debt = _utils.calculateBorrowAmount(_collAmt, _price, COLLATERAL_RATIO * 1000);
         _openTestCDP(_redeemer, _collAmt, _debt);
-        uint _redeemNumber = _utils.generateRandomNumber(1, _cdpNumber - 1, _redeemer);
+        uint256 _redeemNumber = _utils.generateRandomNumber(1, _cdpNumber - 1, _redeemer);
         vm.assume(_redeemNumber > 0);
-        uint _redeemDebt;
-        for (uint i = 0; i < _redeemNumber; ++i) {
+        uint256 _redeemDebt;
+        for (uint256 i = 0; i < _redeemNumber; ++i) {
             CdpState memory _state = _getDebtAndCollShares(_cdpIds[i]);
             _redeemDebt += _state.debt;
             address _owner = sortedCdps.getOwnerAddress(_cdpIds[i]);
-            uint _sugar = eBTCToken.balanceOf(_owner);
+            uint256 _sugar = eBTCToken.balanceOf(_owner);
             vm.prank(_owner);
             eBTCToken.transfer(_redeemer, _sugar);
         }
 
         // execute redemption
-        (bytes32 firstRedempHint, uint partialRedempNICR, , ) = hintHelpers.getRedemptionHints(
+        (bytes32 firstRedempHint, uint256 partialRedempNICR, , ) = hintHelpers.getRedemptionHints(
             _redeemDebt,
             _price,
             0
         );
         require(firstRedempHint == _cdpIds[0], "!firstRedempHint");
-        uint _debtBalBefore = eBTCToken.balanceOf(_redeemer);
+        uint256 _debtBalBefore = eBTCToken.balanceOf(_redeemer);
         vm.prank(_redeemer);
         cdpManager.redeemCollateral(
             _redeemDebt,
@@ -147,11 +147,11 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
             0,
             1e18
         );
-        uint _debtBalAfter = eBTCToken.balanceOf(_redeemer);
+        uint256 _debtBalAfter = eBTCToken.balanceOf(_redeemer);
 
         // post checks
         require(_debtBalAfter + _redeemDebt == _debtBalBefore, "!redemption debt reduction");
-        for (uint i = 0; i < _redeemNumber; ++i) {
+        for (uint256 i = 0; i < _redeemNumber; ++i) {
             require(
                 cdpManager.getCdpStatus(_cdpIds[i]) == 4,
                 "redemption leaves CDP not closed with correct status"
@@ -168,13 +168,13 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         _ensureSystemInvariants();
     }
 
-    function _decMul(uint x, uint y) internal pure returns (uint decProd) {
-        uint prod_xy = x * y;
+    function _decMul(uint256 x, uint256 y) internal pure returns (uint256 decProd) {
+        uint256 prod_xy = x * y;
 
         decProd = (prod_xy + (1e18 / 2)) / 1e18;
     }
 
-    function _decPow(uint _base, uint _minutes) internal pure returns (uint) {
+    function _decPow(uint256 _base, uint256 _minutes) internal pure returns (uint256) {
         if (_minutes > 525600000) {
             _minutes = 525600000;
         } // cap to avoid overflow
@@ -183,9 +183,9 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
             return 1e18;
         }
 
-        uint y = 1e18;
-        uint x = _base;
-        uint n = _minutes;
+        uint256 y = 1e18;
+        uint256 x = _base;
+        uint256 n = _minutes;
 
         // Exponentiation-by-squaring
         while (n > 1) {
@@ -205,7 +205,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
 
     function test_ValidRedemptionsRevertWhenPaused() public {
         (address user, bytes32 userCdpId) = _singleCdpRedemptionSetup();
-        uint debt = 1;
+        uint256 debt = 1;
 
         vm.prank(defaultGovernance);
         cdpManager.setRedemptionsPaused(true);
@@ -213,7 +213,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
 
         vm.startPrank(user);
 
-        (bytes32 firstRedemptionHint, uint partialRedemptionHintNICR, , ) = hintHelpers
+        (bytes32 firstRedemptionHint, uint256 partialRedemptionHintNICR, , ) = hintHelpers
             .getRedemptionHints(debt, (priceFeedMock.fetchPrice()), 0);
 
         vm.expectRevert("CdpManager: Redemptions Paused");
@@ -232,7 +232,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
 
     function test_ValidRedemptionNoLongerRevertsWhenUnpausedAfterBeingPaused() public {
         (address user, bytes32 userCdpId) = _singleCdpRedemptionSetup();
-        uint debt = 1;
+        uint256 debt = 1;
 
         vm.startPrank(defaultGovernance);
         cdpManager.setRedemptionsPaused(true);
@@ -242,7 +242,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
 
         vm.startPrank(user);
 
-        (bytes32 firstRedemptionHint, uint partialRedemptionHintNICR, , ) = hintHelpers
+        (bytes32 firstRedemptionHint, uint256 partialRedemptionHintNICR, , ) = hintHelpers
             .getRedemptionHints(debt, (priceFeedMock.fetchPrice()), 0);
         cdpManager.redeemCollateral(
             debt,
@@ -257,7 +257,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         vm.stopPrank();
     }
 
-    function test_SingleRedemptionCollSurplus(uint _toRedeemICR) public {
+    function test_SingleRedemptionCollSurplus(uint256 _toRedeemICR) public {
         // setup healthy whale Cdp
         // set 1 Cdp that is valid to redeem
         // calculate expected collSurplus from redemption of Cdp
@@ -272,16 +272,16 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         vm.assume(_toRedeemICR > cdpManager.MCR());
         vm.assume(_toRedeemICR <= cdpManager.CCR());
 
-        uint _originalPrice = priceFeedMock.fetchPrice();
+        uint256 _originalPrice = priceFeedMock.fetchPrice();
 
         // ensure there is more than one CDP
         _singleCdpSetupWithICR(user, 200e16);
         (, bytes32 userCdpid) = _singleCdpSetupWithICR(user, _toRedeemICR);
-        uint _totalCollBefore = cdpManager.getEntireSystemColl();
-        uint _totalDebtBefore = cdpManager.getEntireSystemDebt();
-        uint _redeemedDebt = cdpManager.getCdpDebt(userCdpid);
-        uint _cdpColl = cdpManager.getCdpCollShares(userCdpid);
-        uint _cdpLiqReward = cdpManager.getCdpLiquidatorRewardShares(userCdpid);
+        uint256 _totalCollBefore = cdpManager.getEntireSystemColl();
+        uint256 _totalDebtBefore = cdpManager.getEntireSystemDebt();
+        uint256 _redeemedDebt = cdpManager.getCdpDebt(userCdpid);
+        uint256 _cdpColl = cdpManager.getCdpCollShares(userCdpid);
+        uint256 _cdpLiqReward = cdpManager.getCdpLiquidatorRewardShares(userCdpid);
 
         // perform redemption
         _performRedemption(user, _redeemedDebt, userCdpid, userCdpid);
@@ -301,7 +301,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         }
     }
 
-    function test_MultipleRedemptionCollSurplus(uint _toRedeemICR) public {
+    function test_MultipleRedemptionCollSurplus(uint256 _toRedeemICR) public {
         // setup healthy whale Cdp
         // set 3 Cdps that are valid to redeem at same ICR, different borrowers
         // calculate expected collSurplus from full redemption of Cdps
@@ -317,21 +317,21 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         vm.assume(_toRedeemICR > cdpManager.MCR());
         vm.assume(_toRedeemICR <= cdpManager.CCR());
 
-        uint _originalPrice = priceFeedMock.fetchPrice();
+        uint256 _originalPrice = priceFeedMock.fetchPrice();
 
         // ensure there is more than one CDP
         _singleCdpSetupWithICR(users[0], 200e16);
         (, bytes32 userCdpid1) = _singleCdpSetupWithICR(users[0], _toRedeemICR);
         (, bytes32 userCdpid2) = _singleCdpSetupWithICR(users[1], _toRedeemICR + 2e16);
         (, bytes32 userCdpid3) = _singleCdpSetupWithICR(users[2], _toRedeemICR + 4e16);
-        uint _totalCollBefore = cdpManager.getEntireSystemColl();
-        uint _totalDebtBefore = cdpManager.getEntireSystemDebt();
-        uint _cdpDebt1 = cdpManager.getCdpDebt(userCdpid1);
-        uint _cdpDebt2 = cdpManager.getCdpDebt(userCdpid2);
-        uint _cdpDebt3 = cdpManager.getCdpDebt(userCdpid3);
-        uint _cdpColl1 = cdpManager.getCdpCollShares(userCdpid1);
-        uint _cdpColl2 = cdpManager.getCdpCollShares(userCdpid2);
-        uint _redeemedDebt = _cdpDebt1 + _cdpDebt2 + (_cdpDebt3 / 2);
+        uint256 _totalCollBefore = cdpManager.getEntireSystemColl();
+        uint256 _totalDebtBefore = cdpManager.getEntireSystemDebt();
+        uint256 _cdpDebt1 = cdpManager.getCdpDebt(userCdpid1);
+        uint256 _cdpDebt2 = cdpManager.getCdpDebt(userCdpid2);
+        uint256 _cdpDebt3 = cdpManager.getCdpDebt(userCdpid3);
+        uint256 _cdpColl1 = cdpManager.getCdpCollShares(userCdpid1);
+        uint256 _cdpColl2 = cdpManager.getCdpCollShares(userCdpid2);
+        uint256 _redeemedDebt = _cdpDebt1 + _cdpDebt2 + (_cdpDebt3 / 2);
         deal(address(eBTCToken), users[0], _redeemedDebt); // sugardaddy redeemer
 
         // perform redemption
@@ -358,7 +358,7 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
     }
 
     function _singleCdpRedemptionSetup() internal returns (address user, bytes32 userCdpId) {
-        uint debt = 2e17;
+        uint256 debt = 2e17;
         user = _utils.getNextUserAddress();
         userCdpId = _openTestCDP(user, 10000 ether, debt);
 
@@ -367,23 +367,23 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
         vm.stopPrank();
     }
 
-    function _singleCdpSetupWithICR(address _usr, uint _icr) internal returns (address, bytes32) {
-        uint _price = priceFeedMock.fetchPrice();
-        uint _coll = cdpManager.MIN_NET_COLL() * 2;
-        uint _debt = (_coll * _price) / _icr;
+    function _singleCdpSetupWithICR(address _usr, uint256 _icr) internal returns (address, bytes32) {
+        uint256 _price = priceFeedMock.fetchPrice();
+        uint256 _coll = cdpManager.MIN_NET_COLL() * 2;
+        uint256 _debt = (_coll * _price) / _icr;
         bytes32 _cdpId = _openTestCDP(_usr, _coll + cdpManager.LIQUIDATOR_REWARD(), _debt);
-        uint _cdpICR = cdpManager.getICR(_cdpId, _price);
+        uint256 _cdpICR = cdpManager.getICR(_cdpId, _price);
         _utils.assertApproximateEq(_icr, _cdpICR, ICR_COMPARE_TOLERANCE); // in the scale of 1e18
         return (_usr, _cdpId);
     }
 
     function _performRedemption(
         address _redeemer,
-        uint _redeemedDebt,
+        uint256 _redeemedDebt,
         bytes32 _upperPartialRedemptionHint,
         bytes32 _lowerPartialRedemptionHint
     ) internal {
-        (bytes32 firstRedemptionHint, uint partialRedemptionHintNICR, , ) = hintHelpers
+        (bytes32 firstRedemptionHint, uint256 partialRedemptionHintNICR, , ) = hintHelpers
             .getRedemptionHints(_redeemedDebt, priceFeedMock.fetchPrice(), 0);
         vm.prank(_redeemer);
         cdpManager.redeemCollateral(
@@ -400,10 +400,10 @@ contract CDPManagerRedemptionsTest is eBTCBaseInvariants {
     function _checkFullyRedeemedCdp(
         bytes32 _cdpId,
         address _cdpOwner,
-        uint _cdpColl,
-        uint _cdpDebt
+        uint256 _cdpColl,
+        uint256 _cdpDebt
     ) internal {
-        uint _expectedCollSurplus = _cdpColl +
+        uint256 _expectedCollSurplus = _cdpColl +
             cdpManager.LIQUIDATOR_REWARD() -
             ((_cdpDebt * 1e18) / priceFeedMock.fetchPrice());
         assertTrue(sortedCdps.contains(_cdpId) == false);
